Discord: split discord search into own module
Will also be used for "Send to" on the invite page.
