+++ html/admin.html
@@ -331,8 +331,8 @@
         {{ if .discord_enabled }}
         <div id="modal-discord" class="modal">
             <div class="modal-content card">
-                <span class="heading mb-1">{{ .strings.linkDiscord }}<span class="modal-close">&times;</span></span>
-                <p class="content mb-1">{{ .strings.searchDiscordUser }}</p>
+                <span class="heading mb-1"><span id="discord-header"></span><span class="modal-close">&times;</span></span>
+                <p class="content mb-1" id="discord-description"></p>
                 <div class="row">
                     <input type="search" class="col sm field ~neutral !normal input" id="discord-search" placeholder="user#1234">
                 </div>
